feat!: add task providers
BREAKING CHANGE: cmake tools now required for commands
