+++ src/commands/ValgrindCommandProvider.ts
@@ -1,8 +1,9 @@
-import { ValgrindTaskProvider, ValgrindTaskType } from '../tasks/ValgrindTaskProvider';
-import { commands, Task, tasks, window } from 'vscode';
+import { Task, tasks, window } from 'vscode';
+import { ValgrindTaskType } from '../tasks/AbstractValgrindTaskProvider';
+import { ValgrindDebugTaskProvider } from '../tasks/ValgrindDebugTaskProvider';
 
 export class ValgrindCommandProvider {
-  constructor(private taskProvider: ValgrindTaskProvider) {}
+  constructor(private taskProvider: ValgrindDebugTaskProvider) {}
 
   public async valgrindPid(context?: string[], target?: string): Promise<string | undefined> {
     const task = await this.getValgrindTask(target);
@@ -15,15 +16,16 @@ export class ValgrindCommandProvider {
 
   public async valgrindGdbArg(context: string[], target?: string): Promise<string | undefined> {
     const pid = await this.valgrindPid();
-    if (pid) return `target remote | /usr/bin/vgdb --pid=${pid}`;
+    if (pid) return `target remote | vgdb --pid=${pid}`;
   }
 
   private async getValgrindTask(target?: string): Promise<Task | undefined> {
     const taskList = await tasks.fetchTasks({ type: ValgrindTaskType.debugTask });
 
     if (!target) {
-      if (taskList.length == 1) target = <string>taskList[0].definition.target;
-      else {
+      if (taskList.length === 1) {
+        target = <string>taskList[0].definition.target;
+      } else {
         target = (
           await window.showQuickPick(
             taskList.map((task) => ({
